requirements.txt
openai>=1.30.0
PyYAML>=6.0.1
jsonschema>=4.22.0
pytest>=8.2.0
